Splitting logic into smaller files
