+++ .env
@@ -4,4 +4,4 @@ ACCESS_CONTROL_ALLOW_ORIGIN="http://localhost:5173"
 NODE_ENV="local-dev"
 MEMBER_PUBLIC_ID_COOKIE="best-shot-auth"
 JWT_SECRET="asdasdsalikldosaj"
-API_PROVIDER="ge"
+API_PROVIDER="sofa"
