+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "api-best-shot",
-  "version": "2.9.0",
+  "version": "2.10.0",
   "description": "API - Best Shot",
   "main": "index.js",
   "scripts": {
