--- src/domains/data-providers/globo-esporte/metadata.ts
@@ -1,8 +1,0 @@
-export const GLOBO_ESPORTE_MATCHES_API =
-  'https://api.globoesporte.globo.com/tabela/:external_id/fase/:mode-:slug/rodada/:round/jogos/';
-
-export const GLOBO_ESPORTE_TOURNAMENT_API =
-  'https://api.globoesporte.globo.com/tabela/:external_id/classificacao/';
-
-export const GLOBO_ESPORT_TOURNAMENT_STANDINGS_API =
-  'https://api.globoesporte.globo.com/tabela/:external_id/classificacao/';
