+++ src/domains/data-providers/globo-esporte/team.ts
@@ -1,0 +1,15 @@
+import { AppStandingsTeam } from '../typing';
+import { GloboEsporteStandings } from './typing';
+
+type Team = GloboEsporteStandings['classificacao'][number];
+
+export const teamProvider = {
+  parseFromStandings: (team: Team) =>
+    ({
+      externalId: String(team?.equipe_id),
+      matches: team.vitorias,
+      position: team.ordem,
+      wins: team.vitorias,
+      points: team.pontos,
+    } satisfies AppStandingsTeam),
+};
