+++ src/domains/data-providers/globo-esporte/typing.ts
@@ -1,12 +1,22 @@
 export type GloboEsporteApiRound = GloboEsporteApiMatch[];
 
-// Deprecated
+export const GLOBO_ESPORTE_MATCHES_API =
+  'https://api.globoesporte.globo.com/tabela/:external_id/fase/:mode-:slug/rodada/:round/jogos/';
+
+export const GLOBO_ESPORTE_TOURNAMENT_API =
+  'https://api.globoesporte.globo.com/tabela/:external_id/classificacao/';
+
+export const GLOBO_ESPORT_TOURNAMENT_STANDINGS_API =
+  'https://api.globoesporte.globo.com/tabela/:external_id/classificacao/';
+
 export type GloboEsporteApiMatch = {
   id: number;
-  data_realizacao: string;
-  hora_realizacao: string;
+  data_realizacao: string | null;
+  hora_realizacao: string | null;
   placar_oficial_visitante: number | null;
   placar_oficial_mandante: number | null;
+  placar_penaltis_visitante: number | null;
+  placar_penaltis_mandante: number | null;
   equipes: {
     mandante: {
       id: number;
@@ -24,7 +34,8 @@ export type GloboEsporteApiMatch = {
   sede: {
     nome_popular: string;
   };
-  jogo_ja_comecou: boolean;
+  transmissao: boolean | null;
+  jogo_ja_comecou: boolean | null;
 };
 
 export type GloboEsporteStandings = {
